Move SelectionLevel into its own file
