+++ integration-tests/ownership-access-check/tests/query/all-documents-admin.exotest
@@ -1,6 +1,6 @@
 operation: |
   query {
-    documents { 
+    documents {
       id
       content
       user {
