+++ integration-tests/ownership-access-check/tests/query/all-users-non-admin.exotest
@@ -1,6 +1,6 @@
 operation: |
   query {
-    users(orderBy: {id: ASC}) { 
+    users(orderBy: {id: ASC}) {
       id
       name
       documents(orderBy: {id: ASC}) {
